script added to Commands
